Reorganized and updated routers and controllers
